Add files via upload
I changed it so the make file works. The teacher said that the valgrind memmory leaks and the gcc warnings are similar in the sense that this course doesn't go deep enough to require all warnings and memmory leaks to be fixed. As long as nodes are being freed properly, memory leak problems should not be a problem grade wise.